+++ packages/config/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 2.2.7
+
+### Patch Changes
+
+- @latticexyz/common@2.2.7
+- @latticexyz/schema-type@2.2.7
+
 ## 2.2.6
 
 ### Patch Changes