+++ packages/protocol-parser/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @latticexyz/protocol-parser
 
+## 2.2.7
+
+### Patch Changes
+
+- @latticexyz/common@2.2.7
+- @latticexyz/config@2.2.7
+- @latticexyz/schema-type@2.2.7
+
 ## 2.2.6
 
 ### Patch Changes