+++ packages/store-sync/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @latticexyz/store-sync
 
+## 2.2.7
+
+### Patch Changes
+
+- Updated dependencies [a08ba5e]
+  - @latticexyz/store@2.2.7
+  - @latticexyz/stash@2.2.7
+  - @latticexyz/world@2.2.7
+  - @latticexyz/block-logs-stream@2.2.7
+  - @latticexyz/common@2.2.7
+  - @latticexyz/config@2.2.7
+  - @latticexyz/protocol-parser@2.2.7
+  - @latticexyz/recs@2.2.7
+  - @latticexyz/schema-type@2.2.7
+
 ## 2.2.6
 
 ### Patch Changes